Changing Temperature Info from Int to Double

+++ TemperatureInfo.cs
@@ -7,9 +7,9 @@ namespace Azure_Functions
         [JsonProperty(PropertyName = "temp")]
         public double CurrentTemperature { get; set; }
         [JsonProperty(PropertyName = "pressure")]
-        public int Pressure { get; set; }
+        public double Pressure { get; set; }
         [JsonProperty(PropertyName = "humidity")]
-        public int Humidity { get; set; }
+        public double Humidity { get; set; }
         [JsonProperty(PropertyName = "temp_min")]
         public double MinTemperature { get; set; }
         [JsonProperty(PropertyName = "temp_max")]
